% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets_fast.R
\name{get_tweets_fast}
\alias{get_tweets_fast}
\title{Twitter-data retrieval (fast)}
\usage{
get_tweets_fast(username, pages = 1)
}
\arguments{
\item{username}{The name of the twitter follower you wish to lookup.}

\item{pages}{The number of tweet pages you want to retrieve. About 20 tweets per page.}
}
\value{
Returns a tibble containing tweet-data for the specified used.
}
\description{
This function grabs as much data as it can as quickly as it can. This is the fast version of the function which parses the entire tweet page in one go. We are able to do this by ignoring uncertain bits of a tweet. This means skipping things like hashtag extraction, image url extraction and retweet-history.
}
\examples{
get_tweets_fast(username = 'hadleywickham', pages = 5)
}
