% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets_full.R
\name{get_tweets_full}
\alias{get_tweets_full}
\title{Twitter-data retrieval (full)}
\usage{
get_tweets_full(username, pages = 1)
}
\arguments{
\item{username}{The name of the twitter follower you wish to lookup.}

\item{pages}{The number of tweet pages you want to retrieve. About 20 tweets per page.}
}
\value{
Returns a tibble containing tweet-data for the specified used.
}
\description{
This function grabs as much data as it can from Twitter. This is the full version of the function which parses tweets one by one to extract as much info as possible, including hashtags and image links. Column "origins" contains details about where the tweet comes from (i.e. retweet from other users, in chronological order).
}
\examples{
get_tweets_full(username = 'hadleywickham', pages = 5)
}
